Fixed latitude migration

+++ db/migrate/20090401035200_add_extra_fields_to_people.rb
@@ -1,7 +1,7 @@
 class AddExtraFieldsToPeople < ActiveRecord::Migration
   def self.up
-    add_column :people, :latitude, :double
-    add_column :people, :longitude, :double
+    add_column :people, :latitude, :float
+    add_column :people, :longitude, :float
     add_column :people, :image_url, :string
   end
